Bugs crushed.
