+++ src/main/java/chadchat/domain/Message/MessageFactory.java
@@ -1,0 +1,5 @@
+package chadchat.domain.Message;
+
+public interface MessageFactory {
+    Message createMessage(String messageText);
+}
