+++ src/main/java/chadchat/domain/Message/MessageRepository.java
@@ -1,0 +1,5 @@
+package chadchat.domain.Message;
+
+public interface MessageRepository extends MessageFactory{
+    Iterable<Message> findAllMessages();
+}
