--- src/main/java/chadchat/domain/MessageFactory.java
@@ -1,5 +1,0 @@
-package chadchat.domain;
-
-public interface MessageFactory {
-
-}
