--- src/main/java/chadchat/domain/MessageRepository.java
@@ -1,6 +1,0 @@
-package chadchat.domain;
-
-public interface MessageRepository {
-    Iterable<Message> findAllMessages();
-    Message createMessage(Message message);
-}
