--- src/main/java/chadchat/domain/Subscription.java
@@ -1,48 +1,0 @@
-package chadchat.domain;
-
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-
-public class Subscription {
-
-    Channel channel;
-    Boolean subcribe;
-    // UserRights rights;
-    LocalDateTime timestamp;
-
-
-    public Subscription(Channel channel, boolean Subscribe, LocalDateTime timestamp) {
-        // this.message = message;
-        this.channel = channel;
-        // this.rights = rights;
-        this.timestamp = timestamp;
-    }
-
-    /*
-    public static enum UserRights {
-        ALL,
-        LIMITED
-    }
-     */
-
-    public Channel getChannel() {
-        return channel;
-    }
-
-    public void setChannel(Channel channel) {
-        this.channel = channel;
-    }
-
-    public Boolean getSubcribe() {
-        return subcribe;
-    }
-
-    public void setSubcribe(Boolean subcribe) {
-        this.subcribe = subcribe;
-    }
-
-    public LocalDateTime getTimestamp() {
-        return timestamp;
-    }
-}
-
