--- src/main/java/chadchat/infrastructure/CreateUser.java
@@ -1,81 +1,0 @@
-package chadchat.infrastructure;
-
-import chadchat.domain.User;
-
-import java.sql.*;
-import java.util.NoSuchElementException;
-
-import static chadchat.infrastructure.Database.getConnection;
-
-public class CreateUser {
-
-    /*
-
-    public User createUser(String userName) {
-        int id = -1;
-        try (Connection conn = Database.getConnection()) {
-            String sql;
-            sql = "INSERT INTO User(userName) VALUES (?)";
-
-            PreparedStatement stmt = conn.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
-            stmt.setString(1, userName);
-
-
-            stmt.executeUpdate();
-
-            ResultSet rs = stmt.getGeneratedKeys();
-
-            if (rs.next()) {
-                id = rs.getInt(1);
-
-            } else {
-                System.out.println("elsa");
-
-            }
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-
-        }
-        return findUser(id);
-    }
-
-    public User findUser(int id) {
-        try(Connection conn = getConnection()) {
-            PreparedStatement s = conn.prepareStatement(
-                    "SELECT * FROM user WHERE id = ?;");
-            s.setInt(1, id);
-            ResultSet rs = s.executeQuery();
-            if(rs.next()) {
-                return loadUser(rs);
-            } else {
-                System.err.println("No version in properties.");
-                throw new NoSuchElementException("No user with id: " + id);
-            }
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }
-        return null;
-    }
-
-
-    private User loadUser(ResultSet rs) {
-        try {
-            return new User(
-                    rs.getInt("user.id"),
-                    rs.getString("user.userName"),
-                    rs.getTimestamp("user.timestamp").toLocalDateTime());
-            // rs.getBytes("users.salt"),
-            // rs.getBytes("users.secret"));
-        } catch (SQLException e){
-            System.out.println(e.getMessage());
-        }
-        return null;
-    }
-
-    public static void main(String[] args) {
-        System.out.println(new CreateUser().createUser("Janus2"));
-    }
-
-     */
-
-}
